Add FakeInterceptor class to fake responses from local json file

+++ app/src/main/java/com/erkprog/weather/data/entity/ForecastResponse.java
@@ -1,6 +1,7 @@
 package com.erkprog.weather.data.entity;
 
 import java.util.List;
+
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
@@ -313,7 +314,7 @@ public class ForecastResponse {
 
       @SerializedName("Value")
       @Expose
-      private Integer value;
+      private Double value;
       @SerializedName("Unit")
       @Expose
       private String unit;
@@ -321,11 +322,11 @@ public class ForecastResponse {
       @Expose
       private Integer unitType;
 
-      public Integer getValue() {
+      public Double getValue() {
         return value;
       }
 
-      public void setValue(Integer value) {
+      public void setValue(Double value) {
         this.value = value;
       }
 
@@ -352,7 +353,7 @@ public class ForecastResponse {
 
       @SerializedName("Value")
       @Expose
-      private Integer value;
+      private Double value;
       @SerializedName("Unit")
       @Expose
       private String unit;
@@ -360,11 +361,11 @@ public class ForecastResponse {
       @Expose
       private Integer unitType;
 
-      public Integer getValue() {
+      public Double getValue() {
         return value;
       }
 
-      public void setValue(Integer value) {
+      public void setValue(Double value) {
         this.value = value;
       }
